[bug-66215] add test case (that shows we have issues and need fixes)
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1903459 13f79535-47bb-0310-9956-ffa450edef68

+++ poi-ooxml/src/test/java/org/apache/poi/ss/tests/formula/TestStructuredReferences.java
@@ -182,7 +182,7 @@ class TestStructuredReferences {
             //correct all sheet table-reference-formulas which probably got damaged after shift rows
             for (CTTableColumn tableCol : table.getCTTable().getTableColumns().getTableColumnList()) {
                 if (tableCol.getCalculatedColumnFormula() != null) {
-                    int id = (int) tableCol.getId();
+                    int id = Math.toIntExact(tableCol.getId());
                     String formula = tableCol.getCalculatedColumnFormula().getStringValue();
                     int rFirst = table.getStartCellReference().getRow() + table.getHeaderRowCount();
                     int rLast = table.getEndCellReference().getRow() - table.getTotalsRowCount();
